feat: activate deletion chat room
